Gemini/python_backend/requirements.txt
fastapi==0.115.2
uvicorn==0.34.3
python-dotenv==1.0.1
google-adk==1.16.0
pydantic==2.9.0
